Fixed AccumulatedCost and Requests
Requests are being made, but I need to make a change to the database scheme to make sure that requests and offerings are separate.
This will allow a slightly nicer and more flexible interface for information and checking request status.
Accumulated cost has been fixed and will be accessible in bookings table for check out.
Also may build functionality to cancel a request if needed.
Need to remove debug statements later.